Vega v0.3
vega_master.py

it's actually untested as of the moment
should do that soon
